back/pyproject.toml
[tool.poetry]
name = "knowledgebot-backend"
version = "0.1.0"
description = "Backend for KnowledgeBot RAG Platform"
authors = ["404team"]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"

# База данных (PostgreSQL)
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
asyncpg = "^0.29.0"

# Аутентификация (JWT)
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "~3.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.3.4"

# (ОСТАВЛЯЕМ) httpx нужен для ai_client
httpx = "^0.27.0"

# --- УДАЛЕНО ---
# RAG / AI зависимости переехали в 'back-ai'
# chromadb-client = "^0.5.3"
# sentence-transformers = "^3.0.1"
# langchain = "^0.2.7"
# pypdf2 = "^3.0.1"
# python-docx = "^1.1.2"
# --- КОНЕЦ УДАЛЕНИЯ ---


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"